Setup camera2 preview and work on mapping camera2 features to same enum as camera1

+++ goldeneye/src/main/java/co/infinum/goldeneye/models/Size.kt
@@ -18,7 +18,7 @@ data class Size internal constructor(
         return other.height * other.width - width * height
     }
 
-    val aspectRatio = if (this.height != 0 && this.width != 0) this.width / this.height else -1
+    val aspectRatio = if (this.height != 0 && this.width != 0) this.width.toFloat() / this.height.toFloat() else -1f
 }
 
 internal fun Camera.Size.toInternalSize() = Size(width, height)
